requirements.txt
asgiref==3.9.1
diploma-frontend @ file:///home/yahor/Desktop/%D0%90%D1%80%D1%85%D0%B8%D0%B2-%D0%BF%D1%80%D0%BE%D0%B5%D0%BA%D1%82%20%281%29/diploma-frontend/dist/diploma-frontend-0.6.tar.gz#sha256=3a6b1203724b256946919b4d65b362b005ae6a9e1e93d8d765e96aa6a92764ac
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
